api/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2
django-cors-headers==4.3.1
django-redis==5.4.0
django-redis-session-store==0.1.1
django-redis-sessions==0.6.2
djangorestframework==3.14.0
mysqlclient==2.2.4
pytz==2024.1
redis==5.0.3
sqlparse==0.4.4
